Add a networking module for a secure DNS setup
The module uses NetworkManager and systemd-resolved to allow for split
DNS. The global DNS resolver of resolved is a local dnscrypt-proxy2
service. This setup should secure the vast majority of DNS queries by
using DNSCrypt or DoH. Only the queries that are specific to the DNS
server of the link will be routed there and are therefore not encrypted.