Add Sway related modules
This commit adds configurations for Sway, Waybar, swayidle and swaylock.
The configuration consists of a NixOS system part and a home manager
part.

The NixOS system part enables the Sway service which also enables and
configures many other Sway requirements like polkit and OpenGL. Those
requirements need to be configured on the system level. This part of
the configuration also makes sure the requirements for screen sharing
on Sway are met.

The home manager modules contain the user related configuration of Sway,
Waybar, swayidle and swaylock.