Add a workaround for a race condition
There seems to be a race condition between the xdg-desktop-portal DBus
interface startup and the setting of the systemd environment variables.
On slow hardware this leads to a 25 second delay until Waybar shows up.
On better hardware this outcome of the race condition doesn't happen
very often. Therefore the workaround might not be necessary in this
case.

The workaround restarts the xdg-desktop-portal systemd service after
some time to force it to use the new environment variables set by Sway.