Switch to a libvirt user session
Adding a user to the libvirtd group is mostly equivalent to giving the
user root privileges without a password prompt. Nowadays most of the
virtualisation use cases can be achieved with a libvirt user session.
Therefore, this commit removes the libvirtd group from the main user and
instead adds a home manager module to perform some needed configuration.