Add a polkit authentication agent for Sway desktops
This commit uses the agent from the MATE desktop. It is started as part
of the Sway systemd target like other similar services (e.g. swayidle).

With a polkit agent one can use the system connection of libvirt by
providing ones password. This is more secure than assigning the libvirtd
group which is equivalent to passwordless root.