Fix some user space virtualisation issues
This commit allows one to specify the OVMF and swtpm packages to use.
This makes it possible to use dfferent versions for those components in
case there is an issue with their current version.

Furthermore, this commit now also manages some swtpm related
configuration files. Those files contain store paths which need to be
updated accordingly.