Use the Home Manage fzf integration
Formerly fzf was integrated by utilizing the plugin from OMZ. Right now
this plugin doesn't work with fzf versions newer than 0.48.0, because
this version removed the shell scripts in favor of the --zsh cli option
[1]. As a consequence the fzf-share helper script was removed in Home
Manager, too [2].

This commit therefore replaces the old integration with the one provided
by Home Manager. This is cleaner and should always work with the newest
version of fzf.

Links:
  [1] https://github.com/junegunn/fzf/commit/e74b1251c0f579335e03b3e7182cd7a9f88dbe37
  [2] https://github.com/NixOS/nixpkgs/commit/1035a9efdfe4456f224a92ed133cf0605b9eb821