Add a module to use `podman machine` as a secure Docker replacement
The `podman machine` set of commands run a Podman daemon within a
virtual machine. The whole setup runs as a non-privileged user without
any subuid or subgid configuration. The VM creation is triggered after
the user logs in. This is only done if the VM doesn't exist, yet.

This module is also able to create a thin wrapper around the Docker CLI
to start the VM and set the correct DOCKER_HOST environment variable.