Improve Kitty configuration
The issue with not working sudo completion is now fixed. Therefore, the
no-sudo shell integration option is no longer necessary.

So far the cursor was set to a beam by the shell integraiton. Connecting
to a headless machine without the shell integration resulted in the
default cursor, which is a block. To ensure consistency this commit sets
the default cursor to beam and disables the shell integration.